+++ src/PagesLogin/SignIn/index.js
@@ -1,38 +1,37 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { View, Text, StyleSheet, TextInput, TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native'
+import { Ionicons } from '@expo/vector-icons'
 
 import * as Animatable from 'react-native-animatable'
 
 export default function Register() {
     const navigation= useNavigation();
+    const [hidePass, setHidePass] = useState(true);
 
     return(
         <View style={styles.container}>
             <Animatable.View animation="fadeInLeft" delay={500} style={styles.containerHeader}>
-                <Text style={styles.message}> Realize seu cadastro!</Text>
-                <Text style={styles.description}> Complete os campos abaixo </Text>
+                <Text style={styles.message}> Bem-vindo (a)</Text>
             </Animatable.View>
 
             <Animatable.View animation="fadeInUp" style={styles.containerForm}>
-                <Text style={styles.title}> Nome: </Text>
-                <TextInput placeholder="Digite uma senha..." style={styles.input} />
+                <Text style={styles.title}> Email </Text>
+                <TextInput placeholder="Digite seu email..." style={styles.input} />
 
-                <Text style={styles.title}> Email: </Text>
-                <TextInput placeholder="Digite um email..." style={styles.input} />
-
-                <Text style={styles.title}> Senha: </Text>
-                <TextInput placeholder="Digite uma senha..." style={styles.input} />
-
-                <Text style={styles.title}> Confirmar senha: </Text>
-                <TextInput placeholder="Confirme sua senha..." style={styles.input} />
-
-                <TouchableOpacity style={styles.button}>
-                    <Text style={styles.buttonText}>Cadastrar</Text>
+                <Text style={styles.title}> Senha </Text>
+                <View style={styles.inputArea}>
+                    <TextInput placeholder="Digite sua senha..." style={styles.inputPassword} secureTextEntry={hidePass}/>
+                    <TouchableOpacity onPress={ () => setHidePass(!hidePass) }>
+                        <Ionicons name="eye" color="a1a1a1" size={25} />
+                    </TouchableOpacity>
+                </View>
+                <TouchableOpacity style={styles.button} onPress={() => navigation.navigate('Home')}>
+                    <Text style={styles.buttonText}>Acessar</Text>
                 </TouchableOpacity>
 
-                <TouchableOpacity style={styles.buttonRegister} onPress={() => navigation.navigate('SignIn')}>
-                    <Text style={styles.registerText}>Realize seu login aqui </Text>
+                <TouchableOpacity style={styles.buttonRegister} onPress={() => navigation.navigate('Register')}>
+                    <Text style={styles.registerText}>Crie uma conta gratuitamente, clique aqui!</Text>
                 </TouchableOpacity>
 
                 <TouchableOpacity style={styles.buttonRegister} onPress={() => navigation.navigate('ForgetPassword')}>
@@ -76,7 +75,15 @@ const styles = StyleSheet.create({
         borderBottomWidth: 1,
         height: 40,
         marginBottom: 12,
-        fontSize: 16
+        fontSize: 16,
+        width: '95%'
+    },
+    inputPassword:{
+        borderBottomWidth: 1,
+        height: 40,
+        marginBottom: 12,
+        fontSize: 16,
+        width: '90%'
     },
     button:{
         backgroundColor: '#38a69d',
@@ -102,10 +109,8 @@ const styles = StyleSheet.create({
         color:'#1212a1',
         fontSize: 16
     },
-    description:{
-        color: '#dcdcdc',
-        fontSize: 17,
-        fontWeight: 'bold',
-        padding: 4
+    inputArea:{
+        flexDirection: 'row',
+        width: '100%'
     }
 })
